Don't need var
Label: represents a caption in a user interface
Var: an actual variable in a mathematical expression or programming context, an identifier representing a constant, a symbol identifying a physical quantity, a function parameter, or simply a term used as a placeholder in prose

Is var needed? Well, usually in (settings) "var" isn't used, although that's probably because the var element is practically unheard of.
Sigh.